radius_rpc/requirements.txt
amqp==2.6.1
certifi==2020.12.5
chardet==4.0.0
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.0.0
idna==2.10
kombu==4.6.11
mock==4.0.3
nameko==2.13.0
path==15.1.2
path.py==12.5.0
python-dotenv==0.17.1
PyYAML==5.4.1
radius==0.0.4
requests==2.25.1
six==1.15.0
urllib3==1.26.4
vine==1.3.0
Werkzeug==1.0.1
wrapt==1.12.1
